requirements.txt
Django>=3.0
psycopg2>=2.8
djangorestframework>=3.12.4
djangorestframework-simplejwt>=5.2.0
Pillow>=9.0.0
